+++ src/main/java/com/example/Blogify/services/ProfileService.java
@@ -4,8 +4,8 @@ import com.example.Blogify.payloads.ProfileDto;
 
 public interface ProfileService {
 
-    ProfileDto createUserProfile(ProfileDto profileDto, Integer userId);
+    ProfileDto createUserProfile(ProfileDto profileDto, Long userId);
 
-    ProfileDto updateProfile(ProfileDto profileDto, Integer profileId);
+    ProfileDto updateProfile(ProfileDto profileDto, Long profileId);
 
 }
